Respectralized
During the ImHypAd project, we faced issues in the reconstruction of Hyperspectral Images (HSI). One step of the reconstruction using the Regularization Approach (RA) was to regularize the obtained HSI everywhere but on the contour. However, some HSI can have almost similar spectrums, which could occasionally create issues in the contour detection. Hence, changing the spectrums present in the HSI in order to maximize the difference between them can be used in order to facilitate the contour detection. However, the reconstruction will be done on the original HSI (here called IC for Input Cube) while the contour detection will be realized on the modified HSI (here called IC_modified).